app/requirements.txt
cassandra-driver>=3.29.0
venv-pack==0.2.0
pyarrow>=10.0.0
fastparquet>=0.8.3
pathvalidate>=3.0.0
pyspark>=3.3.0
tqdm>=4.64.0
